requirements.txt
asgiref==3.9.1
Django==5.2.6
django-etc==1.4.0
django-hitcount==1.3.5
django-jazzmin==3.0.1
pillow==11.3.0
sqlparse==0.5.3
tzdata==2025.2
